+++ frontend/src/lib/services/auth-service.ts
@@ -1,21 +1,10 @@
 import { goto } from '$app/navigation';
 import { invalidateAll } from '$app/navigation';
-import BaseAPIService from './api-service';
+import BaseAPIService from './api/api-service';
 import userStore from '$lib/stores/user-store';
 import type { User } from '$lib/types/user.type';
-
-export interface LoginCredentials {
-	username: string;
-	password: string;
-}
-
-type LoginResponseData = {
-	token: string;
-	refreshToken: string;
-	expiresAt: string;
-	user: User;
-	requirePasswordChange?: boolean;
-};
+import type { OidcStatusInfo } from '$lib/types/settings.type';
+import type { OidcUserInfo, LoginCredentials, LoginResponseData } from '$lib/types/auth.type';
 
 export class AuthService extends BaseAPIService {
 	async login(credentials: LoginCredentials): Promise<User> {
@@ -38,10 +27,32 @@ export class AuthService extends BaseAPIService {
 
 			return user;
 		} catch {
-			// Clear user from store on error
 			userStore.clearUser();
 			return null;
 		}
+	}
+
+	async getAuthUrl(redirectUri: string): Promise<string> {
+		const response = (await this.handleResponse(this.api.post('/oidc/url', { redirectUri }))) as {
+			authUrl: string;
+		};
+		return response.authUrl;
+	}
+
+	async handleCallback(
+		code: string,
+		state: string
+	): Promise<{
+		success: boolean;
+		token?: string;
+		user?: OidcUserInfo;
+		error?: string;
+	}> {
+		return this.handleResponse(this.api.post('/oidc/callback', { code, state }));
+	}
+
+	async getStatus(): Promise<OidcStatusInfo> {
+		return this.handleResponse(this.api.get('/oidc/status'));
 	}
 }
