+++ frontend/src/lib/services/event-service.ts
@@ -1,8 +1,8 @@
-import BaseAPIService from './api-service';
+import BaseAPIService from './api/api-service';
 import type { SearchPaginationSortRequest, Paginated } from '$lib/types/pagination.type';
 import type { Event } from '$lib/types/event.type';
 
-export default class EventAPIService extends BaseAPIService {
+export default class EventService extends BaseAPIService {
 	async getEvents(options?: SearchPaginationSortRequest): Promise<Paginated<Event>> {
 		const res = await this.api.get('/events', { params: options });
 		return res.data;
@@ -12,3 +12,5 @@ export default class EventAPIService extends BaseAPIService {
 		return this.handleResponse(this.api.delete(`/events/${id}`));
 	}
 }
+
+export const eventService = new EventService();
