+++ frontend/src/routes/auth/login/+page.ts
@@ -7,10 +7,8 @@ export const load = async ({ parent, url }) => {
 		throw redirect(302, '/dashboard');
 	}
 
-	// Get redirect parameter from URL
 	const redirectTo = url.searchParams.get('redirect') || '/dashboard';
 
-	// Get error parameter if any
 	const error = url.searchParams.get('error');
 
 	return {
