+++ frontend/src/routes/auth/logout/+page.ts
@@ -1,5 +1,4 @@
 import { redirect } from '@sveltejs/kit';
-import { browser } from '$app/environment';
 import type { PageLoad } from './$types';
 
 export const load: PageLoad = async ({ fetch }) => {
